Correct -Wformat errors
Fix format string mismatches where the format specifier doesn't match
the argument type:

- Change %d to %u for unsigned int/uint32_t arguments
- Change %u to %d for signed int arguments
- Add casts where needed for printf family functions
- Use PRIu64/PRIi64 for uint64_t/int64_t arguments

Fix -Wformat-nonliteral warnings

Add TIFF_ATTRIBUTE((__format__(__printf__, N, 0))) to functions that
accept format strings as parameters (va_list style). The 0 indicates
no variadic arguments to check since the format is validated at the
call site.

+++ test/raw_decode.c
@@ -114,9 +114,9 @@ static int check_rgba_pixel(int pixel, int min_red, int max_red, int min_green,
     fprintf(stderr,
             "Got R=%d (expected %d..%d), G=%d (expected %d..%d), B=%d "
             "(expected %d..%d), A=%d (expected %d..%d)\n",
-            TIFFGetR(rgba), min_red, max_red, TIFFGetG(rgba), min_green,
-            max_green, TIFFGetB(rgba), min_blue, max_blue, TIFFGetA(rgba),
-            min_alpha, max_alpha);
+            (int)TIFFGetR(rgba), min_red, max_red, (int)TIFFGetG(rgba),
+            min_green, max_green, (int)TIFFGetB(rgba), min_blue, max_blue,
+            (int)TIFFGetA(rgba), min_alpha, max_alpha);
     return 1;
 }
